Raise errors if dpred in `BaseDataMisfit` has nans (#1615)
Make `BaseDataMisfit.residual` to raise an error if the simulation
returns an array with `nan`s and/or `inf`s after calling the `dpred`
method. Add tests to check if the errors are correctly raised. Remove
error in case `BaseDataMisfit.data` is `None`: the `data` setter method
doesn't allow `BaseDataMisfit.data` to be `None`, therefore the error
will never be raised.