Implement `G` matrix as `LinearOperator` in gravity simulation (#1622)
Make the `G` property in the gravity survey to be either a 2D Numpy
array, or a `scipy.sparse.LinearOperator` if `store_sensitivities` is
set to `"forward_only"`. The `LinearOperator` is capable of performing
operations like `G @ m` and `G.T @ v` without allocating the full `G`
matrix on memory or disk. This allows `Jvec` and `Jtvec` to be called
and not allocate the full `G` matrix. The `getJ` method also returns
either a 2D Numpy array or a `LinearOperator` formed by the product
between `G` and the `rhoDeriv` as a `LinearOperator` as well.
Reimplement the `getJtJdiag` method: allow it to get the diagonal
without building `G` if `"forward_only"`, optimize the computation of
the diagonal when `G` is fully formed through `einsum`s, and fix the
cache of `gtg_diagonal` by checking the hash of the `W` argument.
Deprecate the public `gtg_diagonal` property. Extend docstrings of
relevant methods. Add tests for all the new features.

---------

Co-authored-by: Joseph Capriotti <josephrcapriotti@gmail.com>